Initial commit
Some work will be done to clean up the source for this plugin but
functionally it does the job, does it well, and is less prone to
failure than trying to reconstitute a module properly. Instead of
trying to serialize and deserialize the cache, this package writes out
the dependencies, variables, blocks, assets, source and source map of a
module to disk. From those we can create a new HardSourceModule. Like
HMR this requires the use of records. The modules depend on webpack and
loaders being deterministic. Loaders already state that with
`cacheable` and webpack fulfills that through id records.