Extend source map tests and fix full source map support
Source maps in webpack support a wide set of options to customize
output source maps. Mostly for different build time performance
targets. To support them the weirdest thing HardSource needs to do is
recreate the source map options object that webpack creates and doesn't
store in any public location. After that we need to serialize and
expose 4 methods for the different ways source maps are consumed.

- Source.map: returns a rendered source map object
- Source.node: returns an intermediate SourceNode object for generating
  source maps
- Source.listMap: returns an intermediate SourceNode object that
  doesn't care about columns, a easy way to make a source map faster
  while retaining most of its utility
- Module.updateHash: the full chunk source maps (non-eval) include the
  package hash, to recreate that we need to serialize the content built
  into a module's contribution to a chunk's hash