Improve source content key hydration
webpack 1 and 2 depend on different source-map versions which have
different source content name behaviour. To resolve this when
rehydrating the SourceNode for the Source.node method we reach into
RawModule's returned RawSource instance which will come from
webpack-core for webpack 1 and webpack-sources for webpack 2. We can
then get a SourceNode from RawSource and steal SourceNode's
setSourceContent function which will perform the right rehydration
behaviour for the active version of webpack.