Support webpack 2 harmony modules
- Disable tree shaking
- Serialize and deserialize HarmonyExport likeness
- Serialize and deserialize strict module parameter

Tree shaking breaks the assumption this plugin makes. A module only
needs to be rebuilt if it's file dependencies change. Tree shaking
means a module need to be rendered differently because other modules
referring to it have changed what exported members are and are not
used. When serializing we cut out a bunch of information needed for
tree shaking to work since we are looking to be faster and every time a
chunk is rendered that work needs to be reconsidered for tree shaking.