Store assets in a folder
Storing assets in json or text is very wasteful. Store assets in their
native binary (or text if they happen to be text) format. This change
hashes the names (the likely already hashed names) in case a user of
this plugin configures their file-loader (or similar) to store under
folders.