Remove module cache entries that are out of date
When an resolved value in the resolve becomes invalid any hard source
pointing to that is also invalid. We require recordsPath to be set and
ID records persist between builds, and so the IDs cached in the hard
source will continue to point at the now invalid resolution. By busting
that entry in the cache, webpack will build a fresh one pointing to the
new resolution.