When invalidating skip already invalidated modules
Modules are invalidated by removing them from the cache. We currently
don't remove the keys with `delete`. We could visit that but either way
we should make sure the entry in `moduleCache` is an object.

This was causing a bug any time two resolutions were invalidated at the
same time.