Replace isUsed hack with extra compiler run and used invalidation
The isUsed hack disabled tree shaking by forcing all modules to output
all of their exports. Replacing that we need to compare what is used in
a new run on the compiler to an old one. Comparing is the easy part,
finding out the values so we can compare is trickier. The approach here
is to run a second compiler before the main compiler. After it
completes the necessary steps we can compare what attributes changed
that may invalidate a HardModule.

After the pre-emptive compiler invalidates HardModules the original
compiler can run with little extra effort. The original compiler can
make use of a shared cache between the compilers, any new NormalModules
compiled in the pre-emptive compiler can be used here. The original
compiler also has the opportunity to replace any invalidated
HardModules.