Cache directory relative to output relative to cwd
While not recommended its possible to have webpack's `output.path` set
to a relative path, who `cacheDirectory` is relative to. If neither of
these values are absolute, the previous version of the code would
result in an unexpected absolute path from the root of the file system.
That will likely not be what the user wants and will probably need
higher privilege to execute, thus causing an error.