Record the request of specifier imports
Need to record the request of an import specifier that is linked to a
given import. Using the recorded request for a specifier, we create a
hash of imports by request and pull the import for a specifier by
request key. With this we can stand harmony dependencies back up so
that export use tracking can work.