Replace pre-emptive compiler with additional pass
webpack 2 can run as many compile passes as needed by responding to a
new plugin point called need-additional-pass. The way this commit
supports accurate used modules is through using additional passes as
necessary. This has the benefit of only using one pass when only one is
needed and working with other plugins like HMR that the pre-emptive
step was breaking in cases.