Check buildTimestamp before serializing
After the first build of a webpack process in watch mode, modules in
the memory cache get stuck as NormalModules. Since the module
serialization test is so simple, those modules would be re-serialized
until the process was restarted. By adding a check to buildTimestamp we
ca make this a little more like needRebuild. Though in this case it
just makes sure we don't continuously write out cached copies of
modules that haven't changed since their last cached copy.