Add how to work with many config files
Many config files can invalidate the cache by being in a directory and
listed in that option to environmentPaths. Otherwise users need another
package to handle this. HardSource doesn't want to directly support
globing those arguments to simplify the option surface area it needs to
maintain and to not add additional behaviour that users need to
implement if a future version drops direct support for plugins to
determine invalidating the whole cache.