Build a compiler prefix for modules based on subcache name
Most child compilers need to use a subcache so their modules are built
considering their needs. We can reverse engineer a prefix following the
hierarchy of subcaches and the keys pointing to them.

An example of where this is needed is the html plugin and lodash.
Lodash relies on variable modules from webpack. One of those variables
needs a module to perform some utility. Since webpack assigns modules
ids, lodash in the html plugin and lodash used in a normal script built
by webpack will likely need separate modules to account for distinct
module ids of the utility module.