Add loader test with loader testing addDependency
An often used loaderContext method is addDependency. addDependency adds
additional files to track to consider when to invalidate a built module
in the memory cache and in hard source's cache. This means a module
after being transformed by a loader can depend on more than the file
that contains the original source. css preprocessor loaders sass, less,
and stylus use this feature for `@import`'d files as changes to those
files also mean the module that imported them needs to be rebuilt.