Stamp hard source caches with the version of hard-source
In case environmentPaths is disabled or customized the hard source
version is always recorded on the cache so that if the version is
different the cache will be ignored and a fresh build will be created.
This way changes to how the cache is read or written always leads to a
new cache being created and no errors can arise from using a new
version of the plugin with an old version of the cache.