Prefetch hard modules
Like the prefetch plugin, add modules to a compilation before they are
needed by the entry dependency. Since the hard modules can be treated
synchronously we can do this in a pretty quick loop outside of
webpack's normal async flow.

This is fairly repetitive of some of webpack's source, but to achieve
the benefit it gains that is necessary as we need to inject separating
treatment of HardModules to inject them into the compilation faster.