Move module resolve invalidation into the ModuleFactoryPlugin
A more stable resolve invalidation is to wait until the
ModuleFactoryPlugin where it can try to make modules and see if their
identifier is different implying that their id will be different. This
helps support modules that depend on external and delegated modules but
doesn't directly cache and restore those module types.