Add another es2015 export order case
Test that index.js rebuilds when the export changes in other.js and
obj.js. webpack cares about the deepest modules and the order they use
es2015 dependencies. Make sure that when that may change that a change
to the dependency it affects still rebuilds other modules depending on
it.