Rename HarmonyCompatiblilityDependency to Compatibility
webpack@2.2.0 added a HarmonyCompatiblityDependency that we needed to
support for the exports -> __webpack_exports__ module iife wrapper. But
the dependency's name was typo'd so we use the typo'd name to be true
to webpack at the time. webpack has since released a fix for the typo
and hard-source needs to update to that name.