Store the resolved module identifier earlier
To help support the ModuleConcatenationPlugin store the resolved module
identifier on the seal plugin hook. Then when serializing the modules
use that value that was stored earlier. This way when the modules are
normally created and checked by hard-source it will compare against the
original module identifier and not the later concatenated module.