Freeze and thaw modules in a ConcatenatedModule
Store the modules in a ConcatenatedModule. Hijack the modules to record
the replacements performed on the modules and store the transformed
source. Store that source onto the frozen module after the initial
freezing. Store the records of the replacements as well. When thawing
make sure the replacements are still correct. If they aren't trigger an
additional build for any modules that require that.