Test that loader use works in watch mode
- If in a watch-run dereference the compiler from the Watching instance
  to get the context when checking if missing paths for a loader now
  exist

Webpack in watch mode creates a Watching instance that is in charge of
starting compiler runs. In this case the 'watch-run' plugin hook is
called instead of the 'run' hook. Instead of getting a compiler the
'watch-run' hook gets the Watching instance and if the compiler is
desired must be dereferenced from the Watching instance.