LICENSE: add title and extension
The title is not legally mandated, but it's convenient for human consumption; it also works as additional metadata, and is part of the recommended license template text (see http://choosealicense.com/licenses/isc/ and https://opensource.org/licenses/isc-license)

The extension is needed to trigger github's soft-wrapping rendering mode. This also allows the title to be highlighted with the appropriate semantic markup (a markdown heading).