Check if cacheItem.sourceMap.map exists
In our test setup we use the webpack.SourceMapDevToolPlugin() and this seems to produce sourceMap items without a 'map' property. The code assumes that the property exists when 'options' are passed to chooseMap(), in our case this results in a runtime error. I've added a simple reference check for the 'map' property, but none for the 'baseMap' property since I assume it is always present?