Naturally thaw NormalModule
Create a NormalModule and set its member from a frozen version using
the serial schema. This module can be unbuilt and rebuilt as it needs.
It is a NormalModule that "was" built and acts like it without any
smoke and mirrors. Its needRebuild is overwritten to take advantage of
file hashes to invalidate with HardSource's strategy.