Allow custom values of info.mode and info.level (#360)
I tried to customize the info.mode and info.level options but failed to do so. The current code evaluates the OR statement as part of the conditional, i.e. this.options.mode = (this.options.mode || process.env.NODE_ENV === 'test') ? 'test' : 'none'; but I thinkg the intend here is this.options.mode = this.options.mode || (process.env.NODE_ENV === 'test' ? 'test' : 'none');