fix: exclude mini-css loader modules
mini-css does not emit assets when built from webpack's memory cache or
hard-source's disk cache. Exclude the modules that lead to the child
compilations that emit the assets so the assets are always emitted.
This has a minor performance change instead of a large performance hit
because the bulk of the css work is still cached outside of these
excluded modules.