fix: memory leak in TransformNormalModuleFactoryPlugin
Problem: plugin wraps createLoaderContext without checking if it was
already done before. As a result, in a watch mode more and more wrappers
get added on top of the function and eat more and more memory. Fixed by
saving original function on module instance itself and wrapping saved
function instead.